Resolve a number of warnings and some failing tests.
Adding the modal_formula and the gui libraries to the documentation. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14413 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/pbespgsolve/ParityGame.h
@@ -32,7 +32,7 @@ namespace mcrl2 { namespace pbes_system { class pbes; } }
 */
 
 /*! Type of priorities in the game */
-typedef uint16_t priority_t;
+typedef size_t priority_t;
 
 /*! The two players in a parity game (Even and Odd) */
 enum player_t { PLAYER_EVEN =  0,  //!< Even (0)
